Rename _schema_retrieval to _schemata_retrieval
